[refactor/#29] 코드 리뷰 반영 - mapper코드 data layer로 이동
