+++ core/network/onboard/build.gradle.kts
@@ -10,8 +10,6 @@ android {
 
 dependencies {
     api(projects.core.network.common)
-    implementation(projects.core.data.onboard.onboardApi)
-    implementation(projects.core.data.test.testApi)
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.bundles.retrofit)
 }
