+++ core/network/user/build.gradle.kts
@@ -10,7 +10,6 @@ android {
 
 dependencies {
     api(projects.core.network.common)
-    implementation(projects.core.data.user.userApi)
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.bundles.retrofit)
 }
